create model: Request, SmsVerifyResponse and User; rename MainActivity to GetPhoneActivity
