+++ app/src/main/AndroidManifest.xml
@@ -11,7 +11,7 @@
             android:theme="@style/AppTheme">
         <activity android:name=".view.VerifySmsActivity">
         </activity>
-        <activity android:name=".MainActivity">
+        <activity android:name=".view.GetPhoneActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN"/>
