+++ app/src/main/java/com/edward/smsotp/model/Request.kt
@@ -1,0 +1,3 @@
+package com.edward.smsotp.model
+
+data class Request(var code: String, var phone: String)
