+++ app/src/main/java/com/edward/smsotp/model/SmsVerifyResponse.kt
@@ -1,0 +1,3 @@
+package com.edward.smsotp.model
+
+data class SmsVerifyResponse(var status: String, var message: String, var Token: String)
