+++ app/src/main/java/com/edward/smsotp/model/User.kt
@@ -1,0 +1,3 @@
+package com.edward.smsotp.model
+
+data class User(var phone: String, var token: String, var username: String )
