+++ app/src/main/java/com/edward/smsotp/view/GetPhoneActivity.kt
@@ -1,19 +1,19 @@
-package com.edward.smsotp
+package com.edward.smsotp.view
 
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import com.edward.smsotp.view.VerifySmsActivity
+import com.edward.smsotp.R
 import kotlinx.android.synthetic.main.activity_main.*
 
-class MainActivity : AppCompatActivity() {
+class GetPhoneActivity : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
 
         ccp_next_btn.setOnClickListener{
-            val smsIntent= Intent(this@MainActivity, VerifySmsActivity::class.java)
+            val smsIntent= Intent(this@GetPhoneActivity, VerifySmsActivity::class.java)
             startActivity(smsIntent)
         }
     }
